Making the integration test uses python 3.7
Tried to use the official python:3.7 images based on debian, it needed further dependencies to make the orchestrator run (GLIBC versions) that I didn't find a good way to install.

In the end relying on pyenv to install a given python version seems like a good idea as it makes it trivial to test against other python versions.